infrastructure/enrollments/requirements.txt
aws-cdk.core~=1.111
aws-cdk.aws-lambda~=1.111
aws-cdk.aws-dynamodb~=1.111
aws-cdk.aws-iam~=1.111
aws-cdk.aws-events~=1.111
aws-cdk.aws_events_targets~=1.111
